Refactored logging setup to directly attach root_log_handler
